--- src/homework2/task3.py
@@ -1,9 +1,0 @@
-st1 = input('Enter "abrakadabra"')
-st = st1.replace(' ', '')
-
-a = st[0]
-for i in st:
-    if i not in a:
-        a = a + i
-
-print(a)
