--- src/homework2/task8.2.py
@@ -1,15 +1,0 @@
-#If we list all the natural numbers below 10 that are multiples of 3 or 5, we get 3, 5, 6 and 9. The sum of these multiples is 23.
-#Finish the solution so that it returns the sum of all the multiples of 3 or 5 below the number passed in. Additionally, if the #number is negative, return 0 (for languages that do have them).
-#Note: If the number is a multiple of both 3 and 5, only count it once.
-
-
-def solution(number):
-    a = 0
-    for i in range(number):
-        if i % 3 == 0 or i % 5 == 0:
-            a += i
-        i += 1
-    return a
-
-
-print(solution(int(input('Enter number here: '))))
